sw: DOCX export: for soft-page-breaks, export continuous section breaks
When a section break is produced based on layout information such as
soft-page-break and follow page style, it's a bad idea to generate a
page break because the break may be in a different position in Word,
particularly if it was inside a fieldmark instruction.

It wouldn't work that well to ignore such page breaks in
MSWordExportBase::NeedTextNodeSplit() because then they would be created
when reaching the next node anyway, via FindPageDesc() fall-back to
layout.

Unfortunately this breaks the test tdf113849_evenAndOddHeaders.odt
which has a page style with follow-page-style on the first page; on
re-import from DOCX, the continuous section is now no longer converted
to a page break, so pages 2-3 have the wrong header/footers... this
seems impossible to fix in general in the import because it doesn't know
whether the continuous section break coincides with a layout page break.

Arguably this worked before mostly by accident? tdf#113849 isn't about
this afaics... Change the test file to have an explicit page break
there, which round-trips as intended. The real fix would be adding
continuous page style change to Writer i guess...

Change-Id: I00ffe3971607c148a7d5c13b89afb936718611c0